% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextFeatures.R
\docType{package}
\name{AutoTimeData}
\alias{AutoTimeData}
\title{Text features computation for NLP, ML and sentiment analysis.}
\description{
This package includes functions to compute text features to be used
for NLP and sentiment analysis with machine learning.
}
