% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{CleanNLPTokens}
\alias{CleanNLPTokens}
\title{CleanNLP Tokens}
\usage{
CleanNLPTokens(text, ids = 1:length(text))
}
\arguments{
\item{text}{Character vector where each element is a different
document.}

\item{ids}{Optional ids for each text document.}
}
\value{
A data.table object containing cleanNLP tokens as a
  data.table object with token words to lower case and original
  token stored in word.orig column.
}
\description{
Computes tokens from text using cleanNLP and Stanford CoreNLP.
}
