% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senti4sd.R
\name{DownloadSenti4SDData}
\alias{DownloadSenti4SDData}
\title{Download Senti4SD data}
\usage{
DownloadSenti4SDData(senti4sd.path = senti4sd.path.default(), update = FALSE)
}
\arguments{
\item{senti4sd.path}{Directory where to store the downloaded file.}

\item{update}{If FALSE, will only download files if they don't
already exists.}
}
\description{
Downloads Senti4SD jar and DSM data files.
}
