% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{GenerateNgrams}
\alias{GenerateNgrams}
\title{Generate ngrams}
\usage{
GenerateNgrams(tokens, ids, n = 3, all.ngrams = FALSE, keep.vocab = FALSE, ...)
}
\arguments{
\item{tokens}{List of character vectors containing tokens for each
document.}

\item{ids}{Ids of the different documents.}

\item{n}{Generate DTMs with values for ngrams going from (1, 1) to
(1, n).}

\item{all.ngrams}{If TRUE, also generates nrgams going from (1, 1)
to (n, n).}

\item{keep.vocab}{If TURE, returns a list containing both the DTM
and the vocabulary used to build it.}

\item{...}{Additional arguments to pass to MakeVocabulary.}
}
\value{
The DTM containing the ngrams or a list containing the DTM
  and the vocabulary used to build it.
}
\description{
Generates different document term matrix for different values of n.
}
\seealso{
MakeVocabulary
}
