% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senti4sd.R
\name{InitJVM}
\alias{InitJVM}
\title{Init JVM}
\usage{
InitJVM(senti4sd.path = senti4sd.path.default(), heap.size = "2048m")
}
\arguments{
\item{senti4sd.path}{Directory where Senti4SD data is stored.}

\item{heap.size}{Maximum heap size for Java.}
}
\description{
Initialize Java Virtual Machine for Senti4SD.
}
\seealso{
DownloadSenti4SDData
}
