% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senti4sd.R
\name{InitSenti4SD}
\alias{InitSenti4SD}
\title{Init Senti4SD}
\usage{
InitSenti4SD(
  senti4sd.path = senti4sd.path.default(),
  heap.size = "2048m",
  update = FALSE
)
}
\arguments{
\item{senti4sd.path}{Directory where Senti4SD data is stored.}

\item{heap.size}{Maximum heap size for Java.}

\item{update}{If FALSE, will only download files if they don't
already exists.}
}
\description{
Initializes Senti4SD by downloading data, initializing the JVM,
loading the DSM and creating the PolarityVector object.
}
\seealso{
InitJVM

LoadDSMSenti4SD

Senti4SDPolarityVector

DownloadSenti4SDData
}
