% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{IsExclamation}
\alias{IsExclamation}
\title{Is exclamation mark.}
\usage{
IsExclamation(words)
}
\arguments{
\item{words}{Character vector of words.}
}
\value{
TRUE if word only contains exclamation marks.
}
\description{
Determine if words are exclamation marks.
}
