% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{IsLaughter}
\alias{IsLaughter}
\title{Is laughter}
\usage{
IsLaughter(words)
}
\arguments{
\item{words}{Character vector of words.}
}
\value{
TRUE if the word is laughter.
}
\description{
Determines if words are laughter.
}
