% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{IsRepeatedMarks}
\alias{IsRepeatedMarks}
\title{Is repeated marks}
\usage{
IsRepeatedMarks(words)
}
\arguments{
\item{words}{Character vector of words.}
}
\value{
TRUE if the word is repeated (interrogation or exclamation) marks.
}
\description{
Determines if words are uppercase words.
}
