% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{IsUppercaseWord}
\alias{IsUppercaseWord}
\title{Is uppercase word}
\usage{
IsUppercaseWord(words)
}
\arguments{
\item{words}{Character vector of words.}
}
\value{
TRUE if the word is all uppercase.
}
\description{
Determines if words are uppercase words.
}
