% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{IsUserMention}
\alias{IsUserMention}
\title{Is user mention}
\usage{
IsUserMention(words)
}
\arguments{
\item{words}{Character vector of words.}
}
\value{
TRUE if the word is a user mention.
}
\description{
Determines if words are uppercase words.
}
