% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword.R
\name{KeywordBasedFeatures}
\alias{KeywordBasedFeatures}
\title{Keyword based features}
\usage{
KeywordBasedFeatures(tokens, negations)
}
\arguments{
\item{tokens}{Token data.table containing ids for document (id),
sentence (sid) and token (tid), token word and original word.}

\item{negations}{Negation word list.}
}
\value{
A data.table with features computed for each document.
}
\description{
Computes keyword based features.
}
\seealso{
CleanNLPTokens
}
