% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{LastToken}
\alias{LastToken}
\title{Last token}
\usage{
LastToken(tokens)
}
\arguments{
\item{tokens}{Token data.table containing ids for document (id),
sentence (sid) and token (tid).}
}
\value{
A subset of the data.table with one line per document
  containing the last token of the document.
}
\description{
Returns the last token for each document.
}
