% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon.R
\name{LexiconBasedFeatures}
\alias{LexiconBasedFeatures}
\title{Lexicon based Features}
\usage{
LexiconBasedFeatures(tokens, lexicons)
}
\arguments{
\item{tokens}{Token data.table containing ids for document (id),
sentence (sid) and token (tid) and token word.}

\item{lexicons}{List of lexicons as data.table object with word and
score columns.}
}
\description{
Computes lexicon based features for a given list of lexicons.
}
\seealso{
ReadLexicons
}
