% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon.R
\name{LexiconFeatures}
\alias{LexiconFeatures}
\title{Lexicon features}
\usage{
LexiconFeatures(tokens, lexicon, prefix = NULL)
}
\arguments{
\item{tokens}{Token data.table containing ids for document (id),
sentence (sid) and token (tid) and token word.}

\item{lexicon}{Lexicon as a data.table with word and score columns.}

\item{prefix}{Optional prefix to add in front of column names.}
}
\value{
A data.table with features computed for each document.
}
\description{
Computes features for a given lexicon.
}
\seealso{
CleanNLPTokens
}
