% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic.R
\name{LoadDSM}
\alias{LoadDSM}
\title{Load DSM}
\usage{
LoadDSM(filename)
}
\arguments{
\item{filename}{Filename to the DSM binary file.}
}
\value{
VectorSpaceModel object.
}
\description{
Load DSM stored in word2vec binary format.
}
