% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senti4sd.R
\name{LoadDSMSenti4SD}
\alias{LoadDSMSenti4SD}
\title{Load Senti4SD DSM}
\usage{
LoadDSMSenti4SD(senti4sd.path = senti4sd.path.default())
}
\arguments{
\item{senti4sd.path}{Directory where Senti4SD data is stored.}
}
\value{
VectorReader Java object resulting from loading the DSM.
}
\description{
Load Senti4SD's Distributional Semantic Model.
}
\seealso{
DownloadSenti4SDData
}
