% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtm.R
\name{MakeDTM}
\alias{MakeDTM}
\title{Make DTM}
\usage{
MakeDTM(tokens, ids, vocabulary)
}
\arguments{
\item{tokens}{List of character vectors containing tokens for each
document.}

\item{ids}{Ids of the different documents.}

\item{vocabulary}{Vocabulary object.}
}
\value{
A dgTMatrix containing the DTM.
}
\description{
Creates a document term matrix with text2vec.
}
\seealso{
MakeVocabulary
}
