% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_var.R
\name{min_var}
\alias{min_var}
\title{Select columns with minimum amount of variance}
\usage{
min_var(x, min = 1)
}
\arguments{
\item{x}{Input data, which should be either a data frame or matrix.}

\item{min}{Minimum amount of variance to require per column.}
}
\value{
Returns data frame (or matrix, depending on input class) with all non-numeric
  columns and only those numeric columns that meet the minimum amount of variance.
}
\description{
Filters numeric columns by requiring a minimum amount of variance
}
\details{
This function omits missing values.
}
