% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic.R
\name{PolarityVectors}
\alias{PolarityVectors}
\title{Polarity Vectors}
\usage{
PolarityVectors(positive, negative, objective, dsm)
}
\arguments{
\item{positive}{Positive words lexicon.}

\item{negative}{Negative words lexicon.}

\item{objective}{Objective words lexicon.}

\item{dsm}{DSM.}
}
\value{
Vector matrix for positive, negative, objective and
  subjective lexicons.
}
\description{
Returns the polarity vectors for positive, negative, objective and
subjective (positive and negative) lexicons.
}
\seealso{
LoadDSM
}
