% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon.R
\name{ReadLexicons}
\alias{ReadLexicons}
\title{Read Lexicons}
\usage{
ReadLexicons(datadir, emoticons = TRUE)
}
\arguments{
\item{datadir}{Path where lexicons are stored.}

\item{emoticons}{If TRUE, includes lexicons based on emoticons.}
}
\description{
Reads lexicons from Senti4SD directory.
}
