% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic.R
\name{SemanticBasedFeatures}
\alias{SemanticBasedFeatures}
\title{Senti4SD Semantic based features}
\usage{
SemanticBasedFeatures(tokens, vectors, dsm)
}
\arguments{
\item{tokens}{Token data.table containing ids for document (id),
token word.}

\item{vectors}{Polarity vectors for positive, negative, objective
and subjective words.}

\item{dsm}{word2vec object representing the Distributional Semantic
Model.}
}
\value{
A data.table with features computed for each document.
}
\description{
Computes semantic based features.
}
\seealso{
LoadDSM

PolarityVectors
}
