% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senti4sd.R
\docType{data}
\name{senti4sd.env}
\alias{senti4sd.env}
\title{Senti4SD environment}
\format{
An object of class \code{environment} of length 0.
}
\usage{
senti4sd.env
}
\description{
Environment containing Java objects needed by Senti4SD. By default
this is empty to NULL and needs to be initialized with
\code{InitSenti4SD}.
}
\seealso{
InitSenti4SD
}
\keyword{datasets}
