% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{senti4sd.gold}
\alias{senti4sd.gold}
\title{Senti4SD Gold Dataset}
\format{
A \code{data.table} object with 4423 rows and 9 columns:
\describe{
  \item{study}{Study that led to the row.}
  \item{label}{Unique label identifying the row.}
  \item{so.id}{StackOverflow id.}
  \item{post.type}{Type of StackOverflow post.}
  \item{text}{Raw text string.}
  \item{final}{Final polarity.}
  \item{r1}{Polarity according to rater 1.}
  \item{r2}{Polarity according to rater 2.}
  \item{r3}{Polarity according to rater 3.}
}
}
\usage{
senti4sd.gold
}
\description{
Senti4SD Gold Dataset from StackOverflow.
}
\keyword{datasets}
