% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{Senti4SDFeatures}
\alias{Senti4SDFeatures}
\title{Senti4SD features}
\usage{
Senti4SDFeatures(text, label = 1:length(text), datadir)
}
\arguments{
\item{text}{Character vector with text documents.}

\item{label}{Labels or ids of the text documents.}

\item{datadir}{Directory where Senti4SD data is stored}
}
\value{
A list containing three matrices containing the semantic
  features, the lexicon-based features and the keyword-based
  features.
}
\description{
Loads Senti4SD data and computes features on given documents.
}
