% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senti4sd.R
\name{Senti4SDGetVector}
\alias{Senti4SDGetVector}
\title{Senti4SD Get Vector}
\usage{
Senti4SDGetVector(vector)
}
\arguments{
\item{vector}{Java vector object.}
}
\value{
Vector as an R object.
}
\description{
Converts a Java vector to R vector.
}
