% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senti4sd.R
\name{Senti4SDPolarityVector}
\alias{Senti4SDPolarityVector}
\title{Senti4SD Polarity Vector}
\usage{
Senti4SDPolarityVector(vector.reader, dimension = 600L)
}
\arguments{
\item{vector.reader}{VectorReader Java object.}

\item{dimension}{Vector dimension to use.}
}
\value{
PolarityVector Java object.
}
\description{
Returns the PolarityVector object for Senti4SD.
}
\seealso{
LoadDSMSenti4SD
}
