% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senti4sd.R
\name{Senti4SDPolarityVectors}
\alias{Senti4SDPolarityVectors}
\title{Senti4SD Polarity Vectors}
\usage{
Senti4SDPolarityVectors()
}
\value{
A list with a positive, negative, subjective, and objective
  vector.
}
\description{
Computes polarity vectors for positive, negative, subjective
(positive or negative) and objective (neutral) words.
}
