% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senti4sd.R
\name{Senti4SDSemanticBasedFeatures}
\alias{Senti4SDSemanticBasedFeatures}
\title{Senti4SD Semantic based features}
\usage{
Senti4SDSemanticBasedFeatures(tokens, vectors = Senti4SDPolarityVectors())
}
\arguments{
\item{tokens}{Token data.table containing ids for document (id),
and token word.}

\item{vectors}{Polarity vectors for positive, negative, objective
and subjective words.}
}
\value{
A data.table with features computed for each document.
}
\description{
Computes semantic based features.
}
\seealso{
InitSenti4SD

Senti4SDPolarityVectors
}
