% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senti4sd.R
\name{Senti4SDStringRepresentation}
\alias{Senti4SDStringRepresentation}
\title{Senti4SD String Representation}
\usage{
Senti4SDStringRepresentation(string)
}
\arguments{
\item{string}{String to vectorize.}
}
\value{
Vector Java object representation of the string.
}
\description{
Computes the vector representation of a string.
}
\seealso{
Senti4SDGetVector
}
