% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textfeatures2.R
\name{textfeatures2}
\alias{textfeatures2}
\title{textfeatures2}
\usage{
textfeatures2(x)
}
\arguments{
\item{x}{Input data. Should be character vector, data frame, or grouped data
frame (grouped_df) with character variable of interest named "text". If
grouped_df is provided, then features will be averaged across all
observations for each group.}
}
\value{
A tibble data frame with extracted features as columns.
}
\description{
Quick and dirty feature extraction from text vector (like textfeatures but
without the sentiment and politeness analysis)
}
