% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{TokenScore}
\alias{TokenScore}
\title{Token Score}
\usage{
TokenScore(tokens, lexicon, ids = unique(tokens$id))
}
\arguments{
\item{tokens}{Token data.table containing document id and word.}

\item{lexicon}{Lexicon as a data.table with word and score columns.}

\item{ids}{List of all document ids.}
}
\value{
data.table resulting of merging tokens and lexicon by
  adding 0 to tokens not found in lexicon. Column id is set as key.
}
\description{
Computes lexicon score for a set of tokens
}
