% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word2vec.R
\name{word2vec_obs}
\alias{word2vec_obs}
\title{word2vec observations}
\usage{
word2vec_obs(x, n_vectors = 100, threads = 1)
}
\arguments{
\item{x}{Input text}

\item{n_vectors}{Number of vectors (dimensions) defaults to 100}

\item{threads}{Number of threads to use, defaults to 1.}
}
\value{
Matrix
}
\description{
word2vec observations
}
