% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textfeatures.R
\name{textfeatures}
\alias{textfeatures}
\title{textfeatures}
\usage{
textfeatures(x)
}
\arguments{
\item{x}{Input data. Should be character vector or data frame/grouped data
frame with variable of interest named "text"}
}
\value{
A tibble data frame with extracted features as columns
}
\description{
Extracts features from text vector.
}
