% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textfeatures.R
\name{textfeatures}
\alias{textfeatures}
\title{textfeatures}
\usage{
textfeatures(x)
}
\arguments{
\item{x}{Input data. Should be character vector, data frame, or grouped data
frame (grouped_df) with character variable of interest named "text". If
grouped_df is provided, then features will be averaged across all
observations for each group.}
}
\value{
A tibble data frame with extracted features as columns.
}
\description{
Extracts features from text vector.
}
